Rework project structure so source and config are separate
