+++ .gitignore
@@ -3,9 +3,9 @@
 *.iml
 
 config.json
-config.*.json
-!config.test.json
-!config.example.json
+config/*.json
+!config/test.json
+!config/sample.json
 coverage/
 node_modules/
 staticman_key
