+++ source/controllers/home.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const pkg = require('./../package.json')
+const pkg = require('../../package.json')
 
 module.exports = (req, res) => {
   res.send(`Hello from Staticman version ${pkg.version}!`)
