+++ source/controllers/process.js
@@ -1,11 +1,10 @@
 'use strict'
 
-const path = require('path')
-const config = require(path.join(__dirname, '/../config'))
-const errorHandler = require('../lib/ErrorHandler')
 const reCaptcha = require('express-recaptcha')
-const Staticman = require('../lib/Staticman')
 const universalAnalytics = require('universal-analytics')
+const config = require('../config')
+const errorHandler = require('../lib/ErrorHandler')
+const Staticman = require('../lib/Staticman')
 
 function checkRecaptcha (staticman, req) {
   return new Promise((resolve, reject) => {
