+++ source/lib/ErrorHandler.js
