+++ source/lib/GitHub.js
