+++ source/lib/GitService.js
