+++ source/lib/GitServiceFactory.js
