+++ source/lib/Staticman.js
