+++ source/lib/SubscriptionsManager.js
