+++ source/lib/TypeUtils.js
