+++ test/unit/controllers/handlePR.test.js
@@ -1,13 +1,13 @@
-const helpers = require('./../../helpers')
-const sampleData = require('./../../helpers/sampleData')
-const Review = require('../../../lib/models/Review')
+const helpers = require('../../helpers')
+const sampleData = require('../../helpers/sampleData')
+const Review = require('../../../source/lib/models/Review')
 
 let mockSetConfigPathFn
 let mockProcessMergeFn
 let req
 
 // Mock Staticman module
-jest.mock('../../../lib/Staticman', () => {
+jest.mock('../../../source/lib/Staticman', () => {
   return jest.fn().mockImplementation(() => {
     return {
       setConfigPath: mockSetConfigPathFn,
@@ -51,7 +51,7 @@ describe('HandlePR controller', () => {
     const mockReview = new Review(pr.title, pr.body, 'false', pr.head.ref, pr.base.ref)
     const mockGetReview = jest.fn().mockResolvedValue(mockReview)
 
-    jest.mock('../../../lib/GitHub', () => {
+    jest.mock('../../../source/lib/GitHub', () => {
       return jest.fn().mockImplementation(() => {
         return {
           getReview: mockGetReview
@@ -59,7 +59,7 @@ describe('HandlePR controller', () => {
       })
     })
 
-    const handlePR = require('./../../../controllers/handlePR')
+    const handlePR = require('../../../source/controllers/handlePR')
 
     let response = await handlePR(req.params.repository, pr)
     expect(mockGetReview).toHaveBeenCalledTimes(1)
@@ -92,7 +92,7 @@ describe('HandlePR controller', () => {
       const mockGetReview = jest.fn().mockResolvedValue(mockReview)
       const mockDeleteBranch = jest.fn()
 
-      jest.mock('../../../lib/GitHub', () => {
+      jest.mock('../../../source/lib/GitHub', () => {
         return jest.fn().mockImplementation(() => {
           return {
             getReview: mockGetReview,
@@ -101,7 +101,7 @@ describe('HandlePR controller', () => {
         })
       })
 
-      const handlePR = require('./../../../controllers/handlePR')
+      const handlePR = require('../../../source/controllers/handlePR')
 
       await handlePR(req.params.repository, pr)
       expect(mockGetReview).toHaveBeenCalledTimes(1)
@@ -133,7 +133,7 @@ describe('HandlePR controller', () => {
       const mockGetReview = jest.fn().mockResolvedValue(mockReview)
       const mockDeleteBranch = jest.fn()
 
-      jest.mock('../../../lib/GitHub', () => {
+      jest.mock('../../../source/lib/GitHub', () => {
         return jest.fn().mockImplementation(() => {
           return {
             getReview: mockGetReview,
@@ -148,7 +148,7 @@ describe('HandlePR controller', () => {
         throw errorMessage
       })
 
-      const handlePR = require('./../../../controllers/handlePR')
+      const handlePR = require('../../../source/controllers/handlePR')
 
       expect.assertions(4)
       try {
@@ -188,7 +188,7 @@ describe('HandlePR controller', () => {
       const mockDeleteBranch = jest.fn()
       const mockGetReview = jest.fn().mockResolvedValue(mockReview)
 
-      jest.mock('../../../lib/GitHub', () => {
+      jest.mock('../../../source/lib/GitHub', () => {
         return jest.fn().mockImplementation(() => {
           return {
             deleteBranch: mockDeleteBranch,
@@ -197,7 +197,7 @@ describe('HandlePR controller', () => {
         })
       })
 
-      const handlePR = require('./../../../controllers/handlePR')
+      const handlePR = require('../../../source/controllers/handlePR')
 
       await handlePR(req.params.repository, pr)
       expect(mockGetReview).toHaveBeenCalledTimes(1)
