+++ test/unit/lib/GitLab.test.js
@@ -1,9 +1,9 @@
-const mockHelpers = require('./../../helpers')
-const sampleData = require('./../../helpers/sampleData')
-const User = require('../../../lib/models/User')
+const mockHelpers = require('../../helpers')
+const sampleData = require('../../helpers/sampleData')
+const User = require('../../../source/lib/models/User')
 const yaml = require('js-yaml')
-const GitLab = require('../../../lib/GitLab')
-const config = require('../../../config')
+const GitLab = require('../../../source/lib/GitLab')
+const config = require('../../../source/config')
 
 let req
 
@@ -36,7 +36,7 @@ describe('GitLab interface', () => {
       }
     })
  
-    const GitLab = require('./../../../lib/GitLab')
+    const GitLab = require('./../../../source/lib/GitLab')
     const gitlab = new GitLab(req.params) // eslint-disable-line no-unused-vars
 
     expect(mockConstructor.mock.calls[0][0]).toEqual({
@@ -57,7 +57,7 @@ describe('GitLab interface', () => {
       }
     })
 
-    const GitLab = require('./../../../lib/GitLab')
+    const GitLab = require('./../../../source/lib/GitLab')
 
     const oauthToken = 'test-oauth-token'
     const gitlab = new GitLab(Object.assign({}, req.params, {oauthToken})) // eslint-disable-line no-unused-vars
@@ -94,7 +94,7 @@ describe('GitLab interface', () => {
         }
       })
 
-      const GitLab = require('./../../../lib/GitLab')
+      const GitLab = require('./../../../source/lib/GitLab')
       const gitlab = new GitLab(req.params)
 
       return gitlab.readFile(filePath).then(contents => {
@@ -120,7 +120,7 @@ describe('GitLab interface', () => {
         }
       })
 
-      const GitLab = require('./../../../lib/GitLab')
+      const GitLab = require('./../../../source/lib/GitLab')
       const gitlab = new GitLab(req.params)
 
       return gitlab.readFile(filePath).catch(err => {
@@ -156,7 +156,7 @@ describe('GitLab interface', () => {
         }
       })
 
-      const GitLab = require('./../../../lib/GitLab')
+      const GitLab = require('./../../../source/lib/GitLab')
       const gitlab = new GitLab(req.params)
 
       return gitlab.readFile(filePath).catch(err => {
@@ -187,7 +187,7 @@ describe('GitLab interface', () => {
         }
       })
 
-      const GitLab = require('./../../../lib/GitLab')
+      const GitLab = require('./../../../source/lib/GitLab')
       const gitlab = new GitLab(req.params)
 
       return gitlab.readFile(filePath).then(contents => {
@@ -218,7 +218,7 @@ describe('GitLab interface', () => {
         }
       })
 
-      const GitLab = require('./../../../lib/GitLab')
+      const GitLab = require('./../../../source/lib/GitLab')
       const gitlab = new GitLab(req.params)
 
       return gitlab.readFile(filePath, true).then(response => {
@@ -246,7 +246,7 @@ describe('GitLab interface', () => {
         }
       })
 
-      const GitLab = require('./../../../lib/GitLab')
+      const GitLab = require('./../../../source/lib/GitLab')
       const gitlab = new GitLab(req.params)
 
       return gitlab.readFile(filePath).then(contents => {
@@ -274,7 +274,7 @@ describe('GitLab interface', () => {
         }
       })
 
-      const GitLab = require('./../../../lib/GitLab')
+      const GitLab = require('./../../../source/lib/GitLab')
       const gitlab = new GitLab(req.params)
 
       return gitlab.readFile(filePath, true).then(response => {
@@ -306,7 +306,7 @@ describe('GitLab interface', () => {
         }
       })
 
-      const GitLab = require('./../../../lib/GitLab')
+      const GitLab = require('./../../../source/lib/GitLab')
       const gitlab = new GitLab(req.params)
 
       return gitlab.writeFile(
@@ -344,7 +344,7 @@ describe('GitLab interface', () => {
           }
         })
 
-        const GitLab = require('./../../../lib/GitLab')
+        const GitLab = require('./../../../source/lib/GitLab')
         const gitlab = new GitLab(req.params)
         const options = {
           content: 'This is a new file',
@@ -381,7 +381,7 @@ describe('GitLab interface', () => {
         }
       })
 
-      const GitLab = require('./../../../lib/GitLab')
+      const GitLab = require('./../../../source/lib/GitLab')
       const gitlab = new GitLab(req.params)
       const options = {
         branch: 'master',
@@ -445,7 +445,7 @@ describe('GitLab interface', () => {
           }
         })
 
-        const GitLab = require('./../../../lib/GitLab')
+        const GitLab = require('./../../../source/lib/GitLab')
         const gitlab = new GitLab(req.params)
 
         return gitlab.writeFileAndSendReview(
@@ -491,7 +491,7 @@ describe('GitLab interface', () => {
         }
       })
 
-      const GitLab = require('./../../../lib/GitLab')
+      const GitLab = require('./../../../source/lib/GitLab')
       const gitlab = new GitLab(req.params)
       const options = {
         commitBody: '',
@@ -535,7 +535,7 @@ describe('GitLab interface', () => {
         }
       })
 
-      const GitLab = require('./../../../lib/GitLab')
+      const GitLab = require('./../../../source/lib/GitLab')
       const gitlab = new GitLab(req.params)
 
       return gitlab.getCurrentUser().then((user) => {
@@ -556,7 +556,7 @@ describe('GitLab interface', () => {
         }
       })
 
-      const GitLab = require('./../../../lib/GitLab')
+      const GitLab = require('./../../../source/lib/GitLab')
       const gitlab = new GitLab(req.params)
 
       return gitlab.getCurrentUser().catch((err) => {
