+++ test/unit/lib/Notification.test.js
@@ -1,6 +1,6 @@
-const config = require('./../../../config')
-const mockHelpers = require('./../../helpers')
-const Notification = require('./../../../lib/Notification')
+const config = require('../../../source/config')
+const mockHelpers = require('../../helpers')
+const Notification = require('../../../source/lib/Notification')
 
 const mockSendFn = jest.fn()
 const mockMessagesFn = jest.fn(() => ({
